Rename: BaseEntity, UserEntity 위치 변경에 따른 Q클래스 import 변경
